fix assetTotal rename userPnL
